third_party/blink/web_tests/fast/canvas/canvas-createImageBitmap-createPattern-expected.txt

Ensure correct behavior of createPattern with ImageBitmaps.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS testBitmap.width is imageWidth
PASS testBitmap.height is imageHeight
PASS testBitmap.width is imageWidth
PASS testBitmap.height is imageHeight
PASS d[0] == 0 is true
PASS d[1] == 0 is true
PASS d[2] == 0 is true
PASS d[3] == 0 is true
PASS d[0] == 0 is true
PASS d[1] == 0 is true
PASS d[2] == 0 is true
PASS d[3] == 0 is true
PASS d[0] == 0 is true
PASS d[1] == 0 is true
PASS d[2] == 0 is true
PASS d[3] == 0 is true
PASS d[0] == 0 is true
PASS d[1] == 0 is true
PASS d[2] == 0 is true
PASS d[3] == 255 is true
PASS d[0] == 0 is true
PASS d[1] == 0 is true
PASS d[2] == 0 is true
PASS d[3] == 0 is true
PASS d[0] == 0 is true
PASS d[1] == 0 is true
PASS d[2] == 0 is true
PASS d[3] == 0 is true
PASS d[0] == 0 is true
PASS d[1] == 0 is true
PASS d[2] == 0 is true
PASS d[3] == 255 is true
PASS d[0] == 0 is true
PASS d[1] == 0 is true
PASS d[2] == 255 is true
PASS d[3] == 255 is true
PASS d[0] == 0 is true
PASS d[1] == 0 is true
PASS d[2] == 0 is true
PASS d[3] == 0 is true
PASS d[0] == 0 is true
PASS d[1] == 0 is true
PASS d[2] == 0 is true
PASS d[3] == 0 is true
PASS d[0] == 0 is true
PASS d[1] == 0 is true
PASS d[2] == 255 is true
PASS d[3] == 255 is true
PASS d[0] == 0 is true
PASS d[1] == 0 is true
PASS d[2] == 0 is true
PASS d[3] == 255 is true
PASS d[0] == 0 is true
PASS d[1] == 0 is true
PASS d[2] == 0 is true
PASS d[3] == 0 is true
PASS d[0] == 0 is true
PASS d[1] == 0 is true
PASS d[2] == 0 is true
PASS d[3] == 0 is true
PASS d[0] == 0 is true
PASS d[1] == 0 is true
PASS d[2] == 0 is true
PASS d[3] == 255 is true
PASS d[0] == 0 is true
PASS d[1] == 0 is true
PASS d[2] == 0 is true
PASS d[3] == 0 is true
PASS d[0] == 0 is true
PASS d[1] == 0 is true
PASS d[2] == 0 is true
PASS d[3] == 0 is true
PASS d[0] == 0 is true
PASS d[1] == 0 is true
PASS d[2] == 0 is true
PASS d[3] == 255 is true
PASS d[0] == 0 is true
PASS d[1] == 0 is true
PASS d[2] == 0 is true
PASS d[3] == 0 is true
PASS d[0] == 0 is true
PASS d[1] == 255 is true
PASS d[2] == 0 is true
PASS d[3] == 255 is true
PASS d[0] == 0 is true
PASS d[1] == 0 is true
PASS d[2] == 0 is true
PASS d[3] == 255 is true
PASS d[0] == 0 is true
PASS d[1] == 0 is true
PASS d[2] == 255 is true
PASS d[3] == 255 is true
PASS d[0] == 0 is true
PASS d[1] == 255 is true
PASS d[2] == 0 is true
PASS d[3] == 255 is true
PASS d[0] == 255 is true
PASS d[1] == 0 is true
PASS d[2] == 0 is true
PASS d[3] == 255 is true
PASS d[0] == 0 is true
PASS d[1] == 0 is true
PASS d[2] == 255 is true
PASS d[3] == 255 is true
PASS d[0] == 0 is true
PASS d[1] == 0 is true
PASS d[2] == 0 is true
PASS d[3] == 255 is true
PASS d[0] == 255 is true
PASS d[1] == 0 is true
PASS d[2] == 0 is true
PASS d[3] == 255 is true
PASS d[0] == 0 is true
PASS d[1] == 255 is true
PASS d[2] == 0 is true
PASS d[3] == 255 is true
PASS d[0] == 0 is true
PASS d[1] == 0 is true
PASS d[2] == 0 is true
PASS d[3] == 255 is true
PASS d[0] == 0 is true
PASS d[1] == 0 is true
PASS d[2] == 0 is true
PASS d[3] == 0 is true
PASS d[0] == 0 is true
PASS d[1] == 255 is true
PASS d[2] == 0 is true
PASS d[3] == 255 is true
PASS d[0] == 0 is true
PASS d[1] == 0 is true
PASS d[2] == 0 is true
PASS d[3] == 0 is true
PASS d[0] == 0 is true
PASS d[1] == 0 is true
PASS d[2] == 0 is true
PASS d[3] == 0 is true
PASS d[0] == 0 is true
PASS d[1] == 255 is true
PASS d[2] == 0 is true
PASS d[3] == 255 is true
PASS d[0] == 0 is true
PASS d[1] == 0 is true
PASS d[2] == 0 is true
PASS d[3] == 0 is true
PASS d[0] == 0 is true
PASS d[1] == 0 is true
PASS d[2] == 0 is true
PASS d[3] == 255 is true
PASS d[0] == 0 is true
PASS d[1] == 255 is true
PASS d[2] == 0 is true
PASS d[3] == 255 is true
PASS d[0] == 255 is true
PASS d[1] == 0 is true
PASS d[2] == 0 is true
PASS d[3] == 255 is true
PASS d[0] == 0 is true
PASS d[1] == 0 is true
PASS d[2] == 0 is true
PASS d[3] == 255 is true
PASS d[0] == 0 is true
PASS d[1] == 0 is true
PASS d[2] == 255 is true
PASS d[3] == 255 is true
PASS d[0] == 255 is true
PASS d[1] == 0 is true
PASS d[2] == 0 is true
PASS d[3] == 255 is true
PASS d[0] == 0 is true
PASS d[1] == 255 is true
PASS d[2] == 0 is true
PASS d[3] == 255 is true
PASS d[0] == 0 is true
PASS d[1] == 0 is true
PASS d[2] == 255 is true
PASS d[3] == 255 is true
PASS d[0] == 0 is true
PASS d[1] == 0 is true
PASS d[2] == 0 is true
PASS d[3] == 255 is true
PASS d[0] == 0 is true
PASS d[1] == 255 is true
PASS d[2] == 0 is true
PASS d[3] == 255 is true
PASS d[0] == 0 is true
PASS d[1] == 0 is true
PASS d[2] == 0 is true
PASS d[3] == 0 is true
PASS d[0] == 0 is true
PASS d[1] == 0 is true
PASS d[2] == 0 is true
PASS d[3] == 255 is true
PASS d[0] == 0 is true
PASS d[1] == 0 is true
PASS d[2] == 0 is true
PASS d[3] == 0 is true
PASS d[0] == 0 is true
PASS d[1] == 0 is true
PASS d[2] == 0 is true
PASS d[3] == 0 is true
PASS d[0] == 0 is true
PASS d[1] == 0 is true
PASS d[2] == 0 is true
PASS d[3] == 255 is true
PASS d[0] == 0 is true
PASS d[1] == 0 is true
PASS d[2] == 0 is true
PASS d[3] == 0 is true
PASS d[0] == 0 is true
PASS d[1] == 0 is true
PASS d[2] == 0 is true
PASS d[3] == 0 is true
PASS d[0] == 0 is true
PASS d[1] == 0 is true
PASS d[2] == 0 is true
PASS d[3] == 255 is true
PASS d[0] == 0 is true
PASS d[1] == 0 is true
PASS d[2] == 255 is true
PASS d[3] == 255 is true
PASS d[0] == 0 is true
PASS d[1] == 0 is true
PASS d[2] == 0 is true
PASS d[3] == 0 is true
PASS d[0] == 0 is true
PASS d[1] == 0 is true
PASS d[2] == 0 is true
PASS d[3] == 0 is true
PASS d[0] == 0 is true
PASS d[1] == 0 is true
PASS d[2] == 255 is true
PASS d[3] == 255 is true
PASS d[0] == 0 is true
PASS d[1] == 0 is true
PASS d[2] == 0 is true
PASS d[3] == 255 is true
PASS d[0] == 0 is true
PASS d[1] == 0 is true
PASS d[2] == 0 is true
PASS d[3] == 0 is true
PASS d[0] == 0 is true
PASS d[1] == 0 is true
PASS d[2] == 0 is true
PASS d[3] == 0 is true
PASS d[0] == 0 is true
PASS d[1] == 0 is true
PASS d[2] == 0 is true
PASS d[3] == 255 is true
PASS d[0] == 0 is true
PASS d[1] == 0 is true
PASS d[2] == 0 is true
PASS d[3] == 0 is true
PASS d[0] == 0 is true
PASS d[1] == 0 is true
PASS d[2] == 0 is true
PASS d[3] == 0 is true
PASS d[0] == 0 is true
PASS d[1] == 0 is true
PASS d[2] == 0 is true
PASS d[3] == 0 is true
PASS successfullyParsed is true

TEST COMPLETE
